Extract admission client and call context into admission/client.rs
